refactor #step2 - 외부로 공개되는 메서드를 인터페이스로 추출합니다
